+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.4.2] WIP
+## [v0.4.2] 2021-10-30
 ### Added
 - `Activty` and friends.
 - `Presence` and friends.